Update the opencore code to the upstream 2.06 release
This matches the v2.06 tag, commit fe8ccd9d9d2826d7748dcc01bd87005f2c729d1b
in the upstream opencore repository. The codecs_v2/audio/gsm_amr directory
is identical to the one in that commit in the upstream project.

This brings no changes to the actual amr code compared to earlier
syncs, only updates the changelog with the release notes for opencore 2.06.