Update the opencore code to the upstream 2.07 release
This matches the v2.07 tag, commit 5357a13ce3abf877a3355e3f3f76e01f5fb3c7de
in the upstream opencore repository. The codecs_v2/audio/gsm_amr directory
is identical to the one in that commit in the upstream project.

This brings no changes to the actual amr code compared to earlier
syncs, only updates the changelog with the release notes for opencore 2.07.